move slack plugins to separate file
